Attempt at stopping emacs from changing directory all the time
I want to choose a directory and stay in it, and open all files relative
to that directory, but emacs works on a "relative to the current buffer"
basis.  This attempts to fix that by modifying functions like
find-file-noselect-1 and set-visited-file-name to restore the directory
after operation.  Sometimes directory changes still sneak in, though --
I think this is just a case of hunting down any other functions which do
it and turning it off.